requirements.txt
Django==2.0.7
Pillow==5.2.0
requests==2.19.1
jdcal==1.4
ipython
readline
